adding a separate file for states
